clean up unused imports
